requirements.txt
flask>=3.1.2
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.2.2
gunicorn>=23.0.0
pillow>=11.3.0
werkzeug>=3.1.3
requests>=2.32.0
python-dotenv>=1.1.1
Flask-Mail>=0.10.0
